Refactoring of common code from PeekLockQueue to a base class.
Partial refactoring for now, will evolve it further.
