+++ src/DurableTask.ServiceFabric/FabricOrchestrationService.cs
@@ -15,11 +15,9 @@ namespace DurableTask.ServiceFabric
 {
     using System;
     using System.Collections.Generic;
-    using System.Collections.Immutable;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
-    using DurableTask.Common;
     using DurableTask.Tracking;
     using Microsoft.ServiceFabric.Data;
